move header files to include/hw/avatar
