+++ hw/avatar/remote_memory.c
@@ -6,6 +6,7 @@
 #include "qemu/osdep.h"
 #include "qemu/error-report.h"
 #include "hw/sysbus.h"
+#include "hw/avatar/remote_memory.h"
 
 #ifdef TARGET_ARM
 #include "target/arm/cpu.h"
@@ -16,44 +17,7 @@
 #define TYPE_AVATAR_RMEMORY "avatar-rmemory"
 #define AVATAR_RMEMORY(obj) OBJECT_CHECK(AvatarRMemoryState, (obj), TYPE_AVATAR_RMEMORY)
 
-enum RemoteMemoryOperation{
-  AVATAR_READ,
-  AVATAR_WRITE,
-};
-
-
-
-typedef struct MemoryForwardReq{
-  uint64_t id;
-  uint64_t pc;
-  uint64_t address;
-  uint64_t value;
-  uint32_t size;
-  enum RemoteMemoryOperation operation;
-
-} MemoryForwardReq;
-
-typedef struct RemoteMemoryResp{
-    uint64_t id;
-    uint64_t value;
-    uint32_t success;
-} RemoteMemoryResp;
-
-typedef struct AvatarRMemoryState {
-    SysBusDevice parent_obj;
-    MemoryRegion iomem;
-    uint64_t address;
-    uint32_t size;
-    uint64_t request_id;
-    char *rx_queue_name;
-    char *tx_queue_name;
-    QemuAvatarMessageQueue *rx_queue;
-    QemuAvatarMessageQueue *tx_queue;
-    qemu_irq irq;
-} AvatarRMemoryState;
-
-
-static uint64_t get_current_pc(void){
+uint64_t get_current_pc(void){
 #ifdef TARGET_ARM
     ARMCPU *cpu = ARM_CPU(qemu_get_cpu(0));
     return cpu->env.regs[15];
