+++ include/hw/avatar/arm_helper.h
